Issue 1130: reloaded state recreates docker container every time
This "fixes" an issue in the docker module on the stable-2.0 branch where (under some combination of circumstances) the net param in the docker module is seen as "default" and shows up out of compliance every time it is run.  The problem is with the lines of code assuming "or bridge".  It is not entirely clear why those were set, as it seems to work just fine 100% of the time by removing the ors on both lines.

If this fix is not the "best" solution, please comment so we can patch this PR and get this into stable-2.0.  It is unclear if this issue is present in stable-2.1 branch as there are other bugs in ansible core 2.1.x that prevent my team from using that for anything yet.